cerbero: Fix mingw toolchain extract failure in some cases
On Windows 10 with "Developer mode" enabled and when using Python 3.8,
we weren't extracting the mingw toolchain correctly due to multiple
Python bugs:

* os.symlink does not translate `/` to `\` and writes out an invalid
  symlink which cannot be deleted by `rm` or `del`. This has been
  fixed by overriding `os.symlink()` in `hacks.py`

* tarfile does not handle FileExistsError from os.symlink correctly
  and tries to write an empty directory (which fails), and then
  silently ignores it. This means that if you extract a new tarball
  that updates a symlink, it will be ignored. This has been fixed by
  overriding `tarfile.os.symlink()` in `hacks.py`

* Due to the above bug, when symlink permissions aren't available,
  tarfile will always write out an empty directory for symlinks
  pointing to a directory. This has been fixed in a more accurate way
  inside `fix_mingw()` now.

Tested the following cases:

Fresh bootstrap on Windows 10 with Python 3.8 and "Developer mode" on
Fresh bootstrap on Windows 10 with Python 3.8 and "Developer mode" off
Fresh bootstrap on Windows 10 with Python 3.5 and "Developer mode" on
Fresh bootstrap on Windows 10 with Python 3.5 and "Developer mode" off

The same as above, but re-bootstrapping (overwrite).

Fixes https://gitlab.freedesktop.org/gstreamer/cerbero/-/issues/260

+++ cerbero/bootstrap/windows.py
@@ -121,15 +121,13 @@ class WindowsBootstrapper(BootstrapperBase):
             os.makedirs(etc_path)
 
     def fix_mingw(self):
-        # Tar does not create correctly the mingw symlink
+        # On Windows, if the user is not allowed to create symbolic links or if
+        # the Python version is older than 3.8, tarfile creates an empty
+        # directory instead of creating a symlink. This affects the `mingw`
+        # dir which is supposed to be a symlink to `usr/x86_64-w64-mingw32`
         sysroot = os.path.join(self.prefix, 'x86_64-w64-mingw32/sysroot')
         mingwdir = os.path.join(sysroot, 'mingw')
-        # The first time the toolchain is extracted it creates a file with
-        # symlink
-        if not os.path.isdir(mingwdir):
-            os.remove(mingwdir)
-        # Otherwise we simply remove the directory and link back the sysroot
-        else:
+        if not os.path.islink(mingwdir) and os.path.isdir(mingwdir):
             shutil.rmtree(mingwdir)
         shell.symlink('usr/x86_64-w64-mingw32', 'mingw', sysroot)
         # In cross-compilation gcc does not create a prefixed cpp
